'deleting and code improvement'

+++ controllers/groups.js
@@ -1,4 +1,4 @@
-const { Groups } = require("../models");
+const { Groups, User } = require("../models");
 const { randomId } = require("../helpers");
 const ctrl = {}
 
@@ -7,10 +7,10 @@ ctrl.add = (req, res) => {
         const id = randomId(16);
         const { admin, members, name } = req.body;
         const group = await Groups.findOne({group_id: id});
-        const member = await Groups.findOne({username: members});
+        const member = await User.findOne({username: members});
         if(group){
             adding()
-        }else if(member){
+        }else if(!member){
             res.json({message: 'That member don\'t exist'})
         }
         else{
@@ -21,7 +21,6 @@ ctrl.add = (req, res) => {
                 group_id: id
             });
             newGroup.save();
-            console.log(newGroup)
             res.json({message: 'group created successfully'})
         }
     }
@@ -39,18 +38,27 @@ ctrl.get = async (req, res) => {
     const { group_id } = req.body
     const group = await Groups.findOne({group_id: group_id});
     if(group) {
-        res.json({groups: group})
+        res.json({group: group})
     }else{
         res.json({error: 'Group not found'})
     }
 }
 
 ctrl.delete = async (req, res) => {
-    const { group_id } = req.body;
+    const { group_id, username } = req.body;
     const group = await Groups.find({group_id});
     if(group) {
-        await Groups.deleteOne({group_id});
-        res.json({message: 'deleted successfully'})
+        if(group.admin == username){
+            await Groups.deleteOne({group_id});
+            res.json({message: 'deleted successfully'})
+        }else{
+
+            const update = await Groups.findOneAndUpdate({group_id: group_id},{
+                members: 'null',
+           });
+           update.save();
+           res.json({message: 'update successfully'})
+        }
     }else{
         res.json({message: 'group not found'})
     }
